Remove reference from io_service parameter
Something very subtle changed in the Boost.Parameter implementation related to
reference handling.  I once tracked it down but didn't record the details.  In
any case, this seems to fix the issue, so I'm going with this (a change left
over from the old me who looked into this long ago), hoping it doesn't break
anything.